requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.1
